Created SettingsTab and SettingsWidget classes
Created SettingsTab class to take a dictionary describing the settings that a given tab should allow access to and generate a tkinter frame containing widgets that allow editing of each setting, as well as extraction of each widget value, and writing back to the config. Created SettingsWidget as a generic getter/setter for a given setting (only specific widgets supported).